+++ src/main/java/net/sf/jabref/logic/cleanup/MoveFieldCleanup.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import net.sf.jabref.logic.util.OptionalUtil;
 import net.sf.jabref.model.FieldChange;
+import net.sf.jabref.model.cleanup.CleanupJob;
 import net.sf.jabref.model.entry.BibEntry;
 
 /**