+++ src/main/java/net/sf/jabref/model/cleanup/FieldFormatterCleanup.java
@@ -1,11 +1,10 @@
-package net.sf.jabref.logic.cleanup;
+package net.sf.jabref.model.cleanup;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
-import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.model.FieldChange;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.event.EntryEventSource;