+++ src/test/java/net/sf/jabref/logic/bibtexkeypattern/MakeLabelWithDatabaseTest.java
@@ -7,7 +7,6 @@ import net.sf.jabref.model.bibtexkeypattern.DatabaseBibtexKeyPattern;
 import net.sf.jabref.model.bibtexkeypattern.GlobalBibtexKeyPattern;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.model.metadata.MetaData;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -17,7 +16,6 @@ import static org.junit.Assert.assertEquals;
 public class MakeLabelWithDatabaseTest {
 
     private BibDatabase database;
-    private MetaData metadata;
     private BibtexKeyPatternPreferences preferences;
     private GlobalBibtexKeyPattern pattern;
     private DatabaseBibtexKeyPattern bibtexKeyPattern;
@@ -26,7 +24,6 @@ public class MakeLabelWithDatabaseTest {
     @Before
     public void setUp() {
         database = new BibDatabase();
-        metadata = new MetaData();
 
         entry = new BibEntry();
         entry.setField("author", "John Doe");
@@ -35,30 +32,29 @@ public class MakeLabelWithDatabaseTest {
         database.insertEntry(entry);
         pattern = new GlobalBibtexKeyPattern(AbstractBibtexKeyPattern.split("[auth][year]"));
         bibtexKeyPattern = new DatabaseBibtexKeyPattern(pattern);
-        metadata.setBibtexKeyPattern(bibtexKeyPattern);
         preferences = new BibtexKeyPatternPreferences("", "", false, true, true, pattern, ',');
     }
 
     @Test
     public void generateDefaultKey() {
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Doe2016"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyAlreadyExistsDuplicatesStartAtA() {
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         BibEntry entry2 = new BibEntry();
         entry2.setField("author", "John Doe");
         entry2.setField("year", "2016");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry2, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry2, preferences);
         assertEquals(Optional.of("Doe2016a"), entry2.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyAlwaysLetter() {
         preferences = new BibtexKeyPatternPreferences("", "", true, true, true, pattern, ',');
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Doe2016a"), entry.getCiteKeyOptional());
     }
 
@@ -66,18 +62,18 @@ public class MakeLabelWithDatabaseTest {
     public void generateDefaultKeyAlwaysLetterAlreadyExistsDuplicatesStartAtB() {
         preferences = new BibtexKeyPatternPreferences("", "", true, true, true, pattern, ',');
 
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         BibEntry entry2 = new BibEntry();
         entry2.setField("author", "John Doe");
         entry2.setField("year", "2016");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry2, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry2, preferences);
         assertEquals(Optional.of("Doe2016b"), entry2.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyStartDuplicatesAtB() {
         preferences = new BibtexKeyPatternPreferences("", "", false, false, true, pattern, ',');
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Doe2016"), entry.getCiteKeyOptional());
     }
 
@@ -85,17 +81,17 @@ public class MakeLabelWithDatabaseTest {
     public void generateDefaultKeyAlreadyExistsDuplicatesStartAtB() {
         preferences = new BibtexKeyPatternPreferences("", "", false, false, true, pattern, ',');
 
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         BibEntry entry2 = new BibEntry();
         entry2.setField("author", "John Doe");
         entry2.setField("year", "2016");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry2, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry2, preferences);
         assertEquals(Optional.of("Doe2016b"), entry2.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyAlreadyExistsManyDuplicates() {
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         BibEntry entry2 = new BibEntry();
         entry2.setField("author", "John Doe");
         entry2.setField("year", "2016");
@@ -106,73 +102,73 @@ public class MakeLabelWithDatabaseTest {
         entry3.setField("year", "2016");
         entry3.setCiteKey(entry.getCiteKeyOptional().get());
         database.insertEntry(entry3);
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry3, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry3, preferences);
         assertEquals(Optional.of("Doe2016a"), entry3.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyFirstTwoAlreadyExists() {
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         BibEntry entry2 = new BibEntry();
         entry2.setField("author", "John Doe");
         entry2.setField("year", "2016");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry2, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry2, preferences);
         database.insertEntry(entry2);
         BibEntry entry3 = new BibEntry();
         entry3.setField("author", "John Doe");
         entry3.setField("year", "2016");
         entry3.setCiteKey(entry.getCiteKeyOptional().get());
         database.insertEntry(entry3);
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry3, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry3, preferences);
         assertEquals(Optional.of("Doe2016b"), entry3.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyLowerModified() {
         bibtexKeyPattern.setDefaultValue("[auth:lower][year]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("doe2016"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyUpperModified() {
         bibtexKeyPattern.setDefaultValue("[auth:upper][year]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("DOE2016"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateDefaultKeyFixedValue() {
         bibtexKeyPattern.setDefaultValue("[auth]Test[year]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("DoeTest2016"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyShortYear() {
         bibtexKeyPattern.setDefaultValue("[shortyear]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("16"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyAuthN() {
         bibtexKeyPattern.setDefaultValue("[auth2]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Do"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyAuthNShortName() {
         bibtexKeyPattern.setDefaultValue("[auth10]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Doe"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyEmptyField() {
         entry = new BibEntry();
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.empty(), entry.getCiteKeyOptional());
     }
 
@@ -180,7 +176,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyEmptyFieldDefaultText() {
         bibtexKeyPattern.setDefaultValue("[author:(No Author Provided)]");
         entry.clearField("author");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("NoAuthorProvided"), entry.getCiteKeyOptional());
     }
 
@@ -188,42 +184,42 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyEmptyFieldColonInDefaultText() {
         bibtexKeyPattern.setDefaultValue("[author:(Problem:No Author Provided)]");
         entry.clearField("author");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Problem:NoAuthorProvided"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitle() {
         bibtexKeyPattern.setDefaultValue("[title]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AnawesomepaperonJabRef"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleAbbr() {
         bibtexKeyPattern.setDefaultValue("[title:abbr]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AapoJ"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyShorttitle() {
         bibtexKeyPattern.setDefaultValue("[shorttitle]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("awesomepaperJabRef"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyVeryshorttitle() {
         bibtexKeyPattern.setDefaultValue("[veryshorttitle]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("awesome"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyShorttitleINI() {
         bibtexKeyPattern.setDefaultValue("[shorttitleINI]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("apJ"), entry.getCiteKeyOptional());
     }
 
@@ -231,7 +227,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthNM() {
         bibtexKeyPattern.setDefaultValue("[auth4_3]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Wond"), entry.getCiteKeyOptional());
     }
 
@@ -239,7 +235,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthNMLargeN() {
         bibtexKeyPattern.setDefaultValue("[auth20_3]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Wonder"), entry.getCiteKeyOptional());
     }
 
@@ -247,7 +243,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthNMLargeM() {
         bibtexKeyPattern.setDefaultValue("[auth2_4]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.empty(), entry.getCiteKeyOptional());
     }
 
@@ -255,7 +251,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthNMLargeMReallyReturnsEmptyString() {
         bibtexKeyPattern.setDefaultValue("[auth2_4][year]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("2016"), entry.getCiteKeyOptional());
     }
 
@@ -264,7 +260,7 @@ public class MakeLabelWithDatabaseTest {
         preferences = new BibtexKeyPatternPreferences("2", "3", false, true, true, pattern, ',');
         bibtexKeyPattern.setDefaultValue("[auth][year]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Doe3016"), entry.getCiteKeyOptional());
     }
 
@@ -272,7 +268,7 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthIni() {
         bibtexKeyPattern.setDefaultValue("[authIni2]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("DS"), entry.getCiteKeyOptional());
     }
 
@@ -280,48 +276,49 @@ public class MakeLabelWithDatabaseTest {
     public void generateKeyAuthIniMany() {
         bibtexKeyPattern.setDefaultValue("[authIni10]");
         entry.setField("author", "John Doe and Donald Smith and Will Wonder");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("DoeSmiWon"), entry.getCiteKeyOptional());
     }
 
+    @Test
     public void generateKeyTitleTitleCase() {
         bibtexKeyPattern.setDefaultValue("[title:title_case]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AnAwesomePaperonJabref"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleCapitalize() {
         bibtexKeyPattern.setDefaultValue("[title:capitalize]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AnAwesomePaperOnJabref"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleSentenceCase() {
         bibtexKeyPattern.setDefaultValue("[title:sentence_case]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Anawesomepaperonjabref"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleTitleCaseAbbr() {
         bibtexKeyPattern.setDefaultValue("[title:title_case:abbr]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AAPoJ"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleCapitalizeAbbr() {
         bibtexKeyPattern.setDefaultValue("[title:capitalize:abbr]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("AAPOJ"), entry.getCiteKeyOptional());
     }
 
     @Test
     public void generateKeyTitleSentenceCaseAbbr() {
         bibtexKeyPattern.setDefaultValue("[title:sentence_case:abbr]");
-        BibtexKeyPatternUtil.makeLabel(metadata, database, entry, preferences);
+        BibtexKeyPatternUtil.makeLabel(bibtexKeyPattern, database, entry, preferences);
         assertEquals(Optional.of("Aapoj"), entry.getCiteKeyOptional());
     }
 }